+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changes
+- `Subscription`, `SubscriptionMap` and `MessageFilter` are now generic over
+  `<M: Message>` instead of just using `Message`.
 
 ## [0.0.6] - 2019-06-07
 ### Changed